rename edit assets
