+++ src/ducks/Watchlists/Widgets/TopPanel/WatchlistContextMenu.js
@@ -6,7 +6,7 @@ import Button from '@santiment-network/ui/Button'
 import Panel from '@santiment-network/ui/Panel/Panel'
 import { normalizeCSV } from '../../utils'
 import Delete from '../../Actions/Delete'
-import Edit from '../../Actions/Edit'
+import Edit from '../../Actions/Edit/EditAssets'
 import WeeklyReport from '../../Actions/WeeklyReport'
 import Copy from '../../Actions/Copy'
 import VisibilityToggle from '../../Actions/ChangeVisibility'
